add crit_load_factor
crit_load_factor can be passed from web interface as list of single
value or list of 8760 values. If a single value is passed then the load
is scaled just for the outage duration.

NOTE: it appears that some logic in create_loads method of library.py
might not define a load_profile (see "#nlaws: why is no load_profile
defined here?").