include maximize mean SOC in REopt objective function (Obj=5)
this is important for resilience analysis, since it can be economically neutral to charge the battery. so if a battery is only used a few times a year for peak shaving, then without this little incentive to charge the battery when it is free, the mean SOC will be relativel low. therefore, when we simulate outages in post processing using the SOC from REopt, the survivability of the system will be relatively low.