pv annual generation restriction includes pv used to charge battery
under net metering scenarios, excluding the pv to battery generation in
this constraint encourages the optimizer to minimize energy costs by
using the pv system to export energy and charge the battery, then use
the battery to meet the site load - which creates a very difficult
problem to solve. In other words, it is trying to size the battery and
PV system that could meet the entire site load, and pay for it by
exporting PV generation. Problem is, it cannot simultaneously
charge/discharge the battery - although that would be the ideal
solution, by-passing the AnnualElecLoad constraint on PV generation
with the battery.