NMIL dat file creation moved to DatFileManager
Another step in dynamically creating dat files that depend on what
Techs are defined in a scenario.

Note that the DAT attribute of the DatFileManager is currently
redundant with the same attribute in the DatLibrary. Once all DAT file
creation is moved to DatFileManager, then DatLibrary (aka Scenario)
will no longer need to keep track of DAT file strings for command line.
The REopt class will use the DatFileManager.DAT attribute to write the
command line.

Also of note, the DatFileManager takes care of defining the PVNM tech
for REopt by deep-copying PV and changing any attributes as necessary.

All tests passing.