BAU cmd line args
create_run_command now uses DatFileManager command line args, which
have a BAU version. The only difference (in the way use it currently)
is the CapCostSegCount. The difference did not raise any errors
previously since mosel will dynamically fill arrays with zeros as
needed (and the zeros in arrays in the BAU case did not affect the
results).