bug fix: bau case infeasible with outage
not sure why this infeasible is not caught in Results.is_optimal.

Problem was that bau and w/system scenarios were sharing
grid_prod_factor and load_list. When an outage is defined the
grid_prod_factor is set to zero during an outage. For bau this makes
the problem infeasible because it has no way to meet the load during
the outage.

Now, bau scenario has  a separate load profile that has zeros for the
outage.