fix handling of Xpress errors
Two key changes:

1. Without the subprocess.PIPE's the stdout and stderr are not passed
via the communicate() function (Mosel does not have a stdout, so we
just capture stderr).  Prior to the changes on lines 20-21 the
communicate() function was returning None, None regardless of errors
from Mosel.

2. If we try to pass back the output_dict with a key of 'error' the
only message we get is 'error is an invalid key for this object' (or
something like that) because 'error' is not part of the RunOutput
django model. Thus, this type of error handling was removed.