improvements to util min charges
rather than require total util charges  >= min charges, now the LCC
includes a min_charge_adder that accounts for the difference between
the min charges and the total util charges.

From some testing, it seems that the implementations are equivalent as
long as it is never worth absorbing the fixed charges. In other words,
for the scenarios that I evaluated, it was never worth building
solar+storage that reduce the utility charges to less than the minimum
charges. However, with this new implementation, REopt does have the
option of absorbing the minimum charges (perhaps future techs/price
decreases will make it worthwhile to absorb minimum charges).