bug fix: TechClassMinSize for BAU creates infeasible constraint
If user passes a min size for PV > 0, then in the BAU case, this line:
```
forall (b in TechClass) do
    sum (t in Tech, s in Seg) dvSystemSize(t, s) *
TechToTechClassMatrix(t,b) >= TechClassMinSize(b)
end-do
```
creates a constraint of `0 >= pv_kw_min`.

This commit fixes the problem using DatFileManager, but might be better
to put a fix into REopt itself. We could make a more logical fix if the
TechToTechClass for Util was used - for some reason the Util Tech is
not part of the Util.TechClass?