fix arbitrary MinChargeAdder choice
For URDB rate 539fb6e9ec4f024bc1dc0579 the minimum monthly charge is greater than all of the other normal charges, which has exposed a weakness in the MinChargeAdder method.

When this situation occurs, it is an arbitrary decision where the minimum charges are place by the solver, and for this rate the solver was placing the min charges in the demand charges and setting the MinChargeAdder to zero.

In order to make this choice non-arbitrary, I placed a factor of 0.999* MinChargeAdder in the objective function, then added the 0.001 balance back into the LCC in the outputs.