improved Exception save to database
since Exception classes are re-raised multiple times by celery, they
are in effect instantiated several times. This means that anything in
the __init__ method is done several times.

This new method of save_to_db makes it more transparent of when the
exception is saved and does not require custom django model handling.