add "local" APP_ENV
For testing and development, also may be used (or some variation) for
new REopt team Linux server.

Note that the settings in bin/worker may be Mac specific (needs tested
in Windows environment). The paths also depend on where Mosel is
installed - so depends on each machine unfortunately.

The bin/worker script is only run if a program that starts the server
via the Procfile is used. Old way of developing with `python manage.py
runserver` plus starting celery and redis-server will still work (see
README.md for more)