new simulated_load endpoint
returns `loads_kw` for built in profiles, as well as `annual_kwh` given
lat, lon, and `doe_reference_name`.

NOTE: this new endpoint is essentially a duplicate of the `annual_kwh`
endpoint (except for `loads_kw` in response). However, the
documentation on developer.nrel.gov includes the `annual_kwh` endpoint
- so I'm leaving it in the API for now.