fix test_custom_critical_load
`test_custom_critical_load` was not testing what it was meant to. `PV`
was being sized based on net metering, and not the outage. By:
- making the outage longer,
- removing battery size constraints,
- setting `net_metering_limit` to zero,
- setting critical_load_pct to zero in second test,
- and aligning the `critical_loads_kw` values with the `loads_kw`,
the test is now testing `critical_loads_kw`.

Also removed `Wind` since it is unnecessary here (and could lead to
later development problems when wind is fleshed out).