fix issue with negative CapCostYInt
when we have existing_pv, the CapCostYInt is negative. And since
binSegChosen can be 1 for PV and PVNM in the same segment (for
example), when dvSystemSize is zero for one or both of PV and PVNM,
then the TotalTechCapCosts can be negative:

TotalTechCapCosts = 0 + (-1234 * 1)

This happens because the TotalTechCapCosts is part of the minimize
objective.

So, we have to add binSingleBasicTech to the TotalTechCapCosts equation
to prevent this "benefit" of negative TotalTechCapCosts occuring.