fix EtaStorOut accounting
Before this fix, if the initial SOC == the final SOC then the sums of
dvElecToStor and dvElecFrom store were equal; in other words,
EtaStorOut was not being accounted for in the energy from the battery
that meets the site load.

This line:
`dvElecFromStor(ts) / EtaStorOut("1S") <=  dvStoredEnergy(ts-1)`
only limits the maximum value of dvElecFromStor, and doesn't account
for discharge losses.

To check this fix for battery only scenarios:
```
sum(LoadProfile) = sum(GridToLoad) + sum(ElecFromStore)
```
and
```
sum(ElecFromStore) = sum(GridToBatt) * EtaStorIn * EtaStorOut.
```

Note that ElecToStore is GridToBatt * EtaStorIn, such that the
progressive efficiency hits can be accounted as reducing amounts in
GridToBatt, ElecToStore, and ElecFromStore.