new output critical_load_series_kw for outage_simulator
needed for outage simulator. prior method of using
`year_one_electric_load_series_kw` X `critical_load_pct` was incorrect
because `year_one_electric_load_series_kw` already had the critical
load spliced in during the outage period. Now we construct an 8760 (or
user uploads) for the critical load, and save it in the outputs to be
used by the outage simulator.

Note that LoadProfile class writes out the csv for
critical_load_series_kw to be later processed in dispatch.py and
results.py (in order to get it to outage_costs.py, which uses the
outage_simulator).