bug fix: outage_simulator pv_kw_ac_hourly used to be for 1kW system
Not sure when this happened, but it used to be the case that
`pv_kw_ac_hourly` was the normalize PV production from PVWatts, such
that the `outage_simulator` also needed `pv_kw` as an input to
determine the actual PV production profile. Somewhere along the way, we
started using the actual PV production (not normalized), but continued
to scale the hourly profile with `pv_kw`.

These changes also use the new API output
`load_profile.critical_load_series_kw` for the `outage_simulator` (and
`outage_costs`).