update requirements.txt, add h5pyd, rm unused packages
# Mac Users!
In order to `pip install` the new h5pyd requirement, you must upgrade
to python 2.7.15 and pip 10. See
https://github.com/HDFGroup/h5pyd/issues/56 for more.

# all platforms
Will we need to upgrade the python version on linux machines? Probably
not, as the issue seems to be related to SSL certificates on mac os
specifically.

# h5pyd and wind toolkit data
h5pyd is for accessing the wind toolkit data, see
https://github.com/NREL/hsds-examples