user_id was getting filled in with error msg due to location
The `super(SubprocessTimeout, self).__init__(task, self.__name__,
run_uuid, message, traceback)` and similar init statements use
locational arguments, and the new `user_id` field in the base Exception
class was not accounted for. Thus, a timeout message, for example, was
`can only concatenate list (not "str") to list` because it the
traceback list was in place of the message string.