move downloading of urdb rate to validator
Improved exception handling: if the URDB came back with a null response
for a given label or rate/utility name combination the API would
respond with an UnexpectedError. Now, the validator tries to download
the URDB rate before kicking off the Celery tasks and responds with
specific messages for bad URDB input values.