Allow deploying other branches to dev (but still only 1 branch).
Since we don't have the DNS setup for multiple subdomains, this still
only allows a single branch to be deployed at a time, but this will
allow the last deploy to "win" so there's a bit more flexibility for
deploying another branch to dev.

Mainly making this change so I can redeploy master to dev to see if the
dev issues are related to changes in the code, or the underlying
servers.