Shift where XPRESS environment variable is set to fix ordering issues.
Despite XPRESS being set, the xpvars.sh script was unsetting it:
https://github.nrel.gov/gist/nmuerdte/3eeabc220adb14d85169d25ab948a33d#file-xpvars-L2
(this appears to be how this file looks on all our installations).

So the issue was due to the ordering in which things were set: since
XPRESS was being set more globally before the execution of the script,
it was then being unset after the xpvars.sh script was sourced. So shift
setting this variable until after the xpvars.sh script is sourced.