fix PBI accounting
we were separating the degradation and discount terms into LevelizationFactorProdIncent and pwf_prod_incent, but to get an accurate annual cashflow the terms cannot be separated. (It is the same as a/x + b/y != (a+b)/(x+y)). The terms are now calculated correctly using the revamped annuity_escalation function. This is a standard way to account for the present worth of an annuity with inflation.