rm levelization_factor from gross load calculation
the load_profile.py calculation of the gross load is the net load + existing PV production.
there is no reason to include the levelization_factor here since we are calculating the gross load in year zero
and not the average gross load over the lifetime of the PV system.

This bug was found working with Joseph Chang on Intelligent Campus analysis where we ran two scenarios:
1. Existing PV accounted for: input the net campus load and the existing PV capacity
    - results in ~150k NPV, primarily in demand charge savings (low elec costs) and no additional PV
2. No existing PV: input gross load constructed from net load plus PVWatts output for existing PV capacity at NREL
    - results in ~2M+ NPV, again primarily in demand charge savings
    - the much higher NPV in this case is due to the gross load being much higher than case 1, since no levelization factor was part of determining the gross load.(i.e. much more load to peak shave for demand charge savings).
    - also the optimal PV system is ~400kW greater than the existing PV system in this case