Bug fix in constraint for chp hourly opex charges
Under this fix, the O&M costs are tracked by hour so that each hour running incurs a cost as a function of size, and any hour in whihc the CHP system is offline contributes zero O&M under this decision variable. Prior to this fix, a small CHP system could incur negative costs when the system is turned off.