mv fuel burn slope & intercept into data_manager
was in urbd_parse because we used to need utility tariff attributes to build the mathematical model arrays - but that is no longer true with the reformulation. Now the building of fuel burning arrays is more clear, and unnecessary dimensions are removed from some arrays (like "Util" tech in fuel burn slope and intercept).