refactor `CurtailmentTiers` -> `SalesTiersBeyondSiteLoad`
We allow users to input a compensation for exported energy beyond the site load, which is what this tier is being used for. It could be considered curtailment when the default value of zero is used for the excess export compensation, but any other time it is not curtailed energy but is exported energy (though the outputs would label it curtailed energy).