Add explicit None check for BuiltInProfile
BuiltInProfile class used for electric and heating loads was checking "if annual_energy" for assigning default building loads

Python handles 0.0 as None for "if object", so this was writing the default reference building energy if 0.0 was input.

Changed to explicitly check for None and pass 0.0 for annual_energy if that was input.