fix load profile definition in chpbeta scenario.py
some old, unnecessary code has persisted in chpbeta that was removed from develop (updating the LoadProfile model now happens once in process_results instead of in both scenario and process_results). @Bill-Becker maybe we can do the same for the `BoilerModel` and other new models that are getting updated in scenario.py and in process_results.py?