update nested_inputs so off-grid defaults do not incorrectly appear for on-grid runs
Previously, the off-grid defaults for sr_required_pct and min_load_met_pct would appear in the inputs nested dict returned from posting to the API, even for on-grid runs. However, these defaults do not get applied to on-grid runs, so this is likely to mislead users. Now, the default for min_load_met will show as 1.0 and sr_required (for load and PV) will show as 0 for on-grid runs.