Increase Julia pod CPU, decrease celery and Julia count for V3
For a REopt run with BAU and Optimal, we only use one celery pod instead of two

Each Julia pod will now do threading to use 2 CPU for parallel BAU and Optimal cases, so we increase the CPU there.

We only need 10 Julia (and 10 celery) pods to maintain current parallel processing of runs.

I kept an extra 5 (for a total of 15) Julia pods for those to process endpoints which are largely served through the HTTP.jl server in the Julia pods.